requirements.txt
# ============================================================
# ЗАВИСИМОСТИ GROUPBUY MINI APP
# ============================================================
#
# Установка: pip install -r requirements.txt
#
# ============================================================

# ==================== WEB FRAMEWORK ====================
# FastAPI — современный асинхронный веб-фреймворк
fastapi==0.109.0

# Uvicorn — ASGI сервер для запуска FastAPI
uvicorn[standard]==0.27.0

# ==================== БАЗА ДАННЫХ ====================
# Supabase Python SDK
supabase==2.12.0

# ==================== ВАЛИДАЦИЯ И НАСТРОЙКИ ====================
# Pydantic — валидация данных (уже в зависимостях FastAPI)
pydantic==2.5.3

# Pydantic Settings — работа с .env файлами
pydantic-settings==2.1.0

# ==================== АВТОРИЗАЦИЯ ====================
# PyJWT — работа с JWT токенами
PyJWT==2.8.0

# Python-jose — альтернатива для JWT (больше возможностей)
python-jose[cryptography]==3.3.0

# Requests — синхронный HTTP клиент (для простых случаев)
requests==2.31.0

# ==================== ПЛАТЕЖИ ====================
# ЮKassa SDK
yookassa==3.1.0

# ==================== РАБОТА С ДАННЫМИ ====================
# Python-dateutil — удобная работа с датами
python-dateutil==2.8.2

# ==================== БЕЗОПАСНОСТЬ ====================
# Passlib — хеширование (если понадобится)
passlib[bcrypt]==1.7.4

# ==================== УТИЛИТЫ ====================
# Python-multipart — для загрузки файлов
python-multipart==0.0.6

# Aiofiles — асинхронная работа с файлами
aiofiles==23.2.1

# ==================== РАЗРАБОТКА ====================
# Pytest — тестирование
pytest==7.4.4

# Pytest-asyncio — асинхронные тесты
pytest-asyncio==0.23.3

# ==================== ОПЦИОНАЛЬНО ====================
# Если понадобится работа с изображениями:
# Pillow==10.2.0

# Если понадобится кэширование:
# redis==5.0.1

# Если понадобится очередь задач:
# celery==5.3.6
# или
# arq==0.25.0

python-dotenv==1.0.0